==> src/03_feature-construction.ipynb <==
import pandas as pd
import numpy as np
import pickle
from sklearn.feature_extraction.text import CountVectorizer

## Load data
df = pd.read_csv("../data/preprocessed_text_w_labels.csv", index_col = 0)

## Get titles
titles = df["Title"].values

## Vectorize
vec = CountVectorizer(binary=True)
vec.fit(titles)
features = vec.transform(titles).toarray()

features.shape

## Save vectorizer and features
np.save("../data/features.npy", features)

pickle.dump(vec, open("../data/vectorizer.p", "wb"))
