==> src/05_compile-baselines.ipynb <==
import os
import json

baseline_results = [os.path.join("../results/", f) for f in os.listdir("../results/") if ".json" in f]

baseline_results

baselines = []
for fname in baseline_results:
    baselines.append(json.load(open(fname, "r")))

def add_dicts(d1, d2):
    for k1 in d1.keys():
        for k2 in d1[k1].keys():
            d1[k1][k2] += d2[k1][k2]
    return d1

for i in range(len(baselines)):
    if i == 0:
        continue
    add_dicts(baselines[0], baselines[i])

baselines[0]

import pandas as pd

baseline_df = pd.DataFrame.from_records(baselines[0]).T

baseline_df.sort_values(by = "accuracy", ascending = False, inplace = True)

baseline_df.round(3).to_csv("../results/baselines.csv")

baseline_df.round(3)

baseline_df.sort_values(by = "time").round(3)
